requirements.txt
Django==4.1
djangorestframework==3.12.4
pytz
beautifulsoup4 
requests
bs4